Led class with a bunch of actions that it can do.
Created a bunch of handlers to manage the internal workings of the class

The example config has 5 buttons on these pins
20, 23, 24, 25, 26
and one led on
21

But for testing you could just put at least 2 buttons on your pi and
assign a "TOGGLE" to one and then each of the other actions one at a time to the second.

Things like "DELAYED_BLINK_OFF" won't work unless they are on to start with.